@@ -7,55 +7,55 @@
   "license": "MIT",
   "private": false,
   "dependencies": {
-    "@mui/material": "^5.2.5",
-    "destyle.css": "^3.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@testing-library/react": "^12.1.4",
+    "jest-transform-css": "^3.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.15.0",
-    "@babel/eslint-parser": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
-    "@techtrain/cli-railway": "^0.1.8",
-    "@types/jest": "^26.0.24",
-    "@types/react": "^17.0.16",
-    "@types/react-dom": "^17.0.9",
+    "@babel/cli": "^7.17.6",
+    "@babel/core": "^7.17.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
+    "@techtrain/cli-railway": "^0.1.11",
+    "@types/jest": "^27.4.1",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-test-renderer": "^17.0.1",
-    "autoprefixer": "^10.3.1",
-    "babel-jest": "^27.0.6",
-    "babel-loader": "^8.2.2",
+    "autoprefixer": "^10.4.4",
+    "babel-jest": "^27.5.1",
+    "babel-loader": "^8.2.4",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "browserslist": "^4.16.7",
+    "browserslist": "^4.20.2",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.2.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-webpack": "^0.13.1",
-    "eslint-plugin-import": "^2.24.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "jest": "^27.0.6",
+    "css-loader": "^6.7.1",
+    "eslint": "^8.12.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-webpack": "^0.13.2",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^27.5.1",
     "jest-css-modules": "^2.1.0",
-    "postcss": "^8.3.6",
-    "postcss-loader": "^6.1.1",
-    "postcss-preset-env": "^6.7.0",
-    "prettier": "^2.3.2",
-    "react-refresh": "^0.10.0",
-    "react-test-renderer": "^17.0.2",
-    "simple-git-hooks": "^2.5.1",
-    "style-loader": "^3.2.1",
-    "ts-jest": "^27.0.4",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
-    "webpack": "^5.49.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "postcss": "^8.4.12",
+    "postcss-loader": "^6.2.1",
+    "postcss-preset-env": "^7.4.3",
+    "prettier": "^2.6.1",
+    "react-refresh": "^0.11.0",
+    "react-test-renderer": "^18.0.0",
+    "simple-git-hooks": "^2.7.0",
+    "style-loader": "^3.3.1",
+    "ts-jest": "^27.1.4",
+    "ts-loader": "^9.2.8",
+    "typescript": "^4.6.3",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   },
   "scripts": {
     "dev": "webpack serve",
