@@ -1,6 +1,7 @@
 import * as React from 'react'
-import * as ReactDOM from 'react-dom'
+import * as ReactDOM from 'react-dom/client'
 import { App } from './App'
 
 const root = document.getElementById('root')
-ReactDOM.render(<App />, root)
+const reactRoot = ReactDOM.createRoot(root)
+reactRoot.render(<App />)
