@@ -1,7 +1,8 @@
 import * as React from 'react'
 import renderer, { act } from 'react-test-renderer'
-import { App } from '../src/App'
 import { fetchMock } from './mock/fetch'
+
+const { App } = require('../src/App') as { App: React.ComponentType<{}> }
 
 describe('Station No.1', () => {
   const fetch = jest.fn()
