@@ -1,7 +1,8 @@
 import * as React from 'react'
-import { App } from '../src/App'
 import { fetchMock } from './mock/fetch'
 import { createAsync } from './utils/createAsync'
+
+const { App } = require('../src/App') as { App: React.ComponentType<{}> }
 
 describe('Station No.3', () => {
   const fetch = jest.fn()
